merge_funcs

+++ labs_god/lab4/merge_sort/merge_funcs.h
@@ -3,7 +3,7 @@
 
 // Прототипы. Реализации находятся в merge_funcs.c
 void copyNumbers(int[], int[], int, int);
-int *mergeArrays(int array1[], int size1, int array2[], int size2, int *compares, int *swaps);
-int *mergeSort(int[], int, int *compares, int *swaps);
+int *mergeArrays(int array1[], int size1, int array2[], int size2, long long *compares, long long *swaps);
+int *mergeSort(int[], int, long long *compares, long long *swaps);
 
 #endif
